% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitJ.R
\name{fitJ}
\alias{fitJ}
\title{Find the maximum-likelihood value of J for a community change dataset}
\usage{
fitJ(
  occs,
  ages,
  sampled = TRUE,
  generationtime = 1,
  CI = FALSE,
  searchinterval = c(1, 9),
  CI.df = 1,
  handle.ext = "condition"
)
}
\arguments{
\item{occs}{matrix of the number of observations in each species at each time,
or else a list containing several such matrices.
In each matrix there is one column per species and one row per time slice.
Time goes from oldest at the bottom to youngest at the top.}

\item{ages}{vector containing the ages of each time slice, in years, from
oldest to youngest, or else a list of such vectors if there are multiple
occurrence matrices. Numbers can run from high to low (e.g., millions of years
ago) or from low to high (e.g., years from the start of a simulation).}

\item{sampled}{boolean indicating whether occs represents a sampled
community (TRUE) or instead represents true species abundances (FALSE).}

\item{generationtime}{time between generations, in years. Default is 1.}

\item{CI}{boolean indicating whether or not to calculate a confidence interval
on the maximum-likelihood estimate of J. Uses function CIfunc_J.}

\item{searchinterval}{consider values of log10J within this interval.}

\item{CI.df}{integer giving the number of degrees of freedom to use when
using the chi-square lookup table to calculate confidence intervals. Should be
1 unless calculating a simultaneous confidence interval.}

\item{handle.ext}{string specifying how extinction (and monodominance) should be
handled. Three options: "ignore" simply passes over these transitions.
"condition" returns logliks conditional on non-extinction and non-monodominance.
"calculate" returns logliks that include species that hit abundance 0 or 1 in
the likelihood calculation. "condition" is the default. Passed to xxprob.}
}
\value{
Returns a list containing "loglik", "J", and (optionally) "CI".
}
\description{
Uses optimize() to fit the J parameter in Hubbell's neutral theory to a dataset.
Optionally, outputs a confidence interval.
}
\details{
optimize() sometimes tests J values which are so unlikely as to return -Inf.
This is fine, and suppressWarnings() just keeps this quiet.
}
\examples{
sim <- simDrift(startingabs=rep(1000,5),ts=c(0,100,200,300,400,500),ss=1000)
fitJ(occs=sim$simulation,ages=sim$times,CI=TRUE)
}
