% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxprob.R
\name{xxprob}
\alias{xxprob}
\title{Probability of a set of transitions under neutral theory}
\usage{
xxprob(
  log10J,
  occs,
  ages,
  sampled = TRUE,
  generationtime = 1,
  ignore.ext = FALSE
)
}
\arguments{
\item{log10J}{log J (community size). Easier for fitJ() to optimize in log space.}

\item{occs}{matrix of the number of observations in each species at each time,
or else a list containing several such matrices.
In each matrix there is one column per species and one row per time slice.
Time goes from oldest at the bottom to youngest at the top.}

\item{ages}{vector containing the ages of each time slice, in years, from
oldest to youngest, or else a list of such vectors if there are multiple
occurrence matrices.}

\item{sampled}{boolean indicating whether occs represents a sampled
community (TRUE) or instead represents true species abundances (FALSE).}

\item{generationtime}{time between generations, in years.}

\item{ignore.ext}{boolean indicating whether transitions that end in 0 should be excluded
from likelihood calculation. FALSE by default. Passed to xprob.}
}
\value{
Returns loglik value.
}
\description{
Gives the likelihood of a community composition timeseries under Hubbell's
neutral theory for a given value of J. Unlike its helper function xprob,
xxprob takes data in the form of the number of individuals observed in each
species, rather than relative abundances. Can take either a single dataset or
a list of them.
}
\details{
Assumes non-overlapping generations. Because the user supplies the ages of each
time slice, the function needs to be given a generation time to know how many
generations separate each adjacent pair of time slices. Default is 1 year.
}
\examples{
mat <- matrix(data=c(52,12,160,109,30,401,93,31,355),nrow=3)
xxprob(log10J=5,occs=mat,ages=c(200,100,0)) #7.257928
}
